requirements.txt
python-telegram-bot[job-queue]==21.11.1
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.76.0
requests==2.32.3
